requirements.txt
flask
pillow
numpy
scikit-learn
joblib
